Move theming to App.js
