+++ src/themes/themes.json
@@ -58,7 +58,7 @@
         "paper": "#A1B57D"
       },
       "text": {
-        "primary": "#fff"
+        "primary": "#101a1f"
       },
       "primary": {
         "main": "#B33030"
